requirements.txt
flask==3.0.2
flask-cors==4.0.0
twilio==8.13.0
supabase==2.4.2
openai==1.12.0
python-dotenv==1.0.1
gunicorn==21.2.0
pydantic==2.6.3
httpx==0.27.0
requests==2.31.0
ffmpeg-python==0.2.0
python-dateutil==2.9.0
apscheduler==3.10.4
reportlab==4.1.0
Pillow==10.2.0
